Some refactoring due to code review
